+++ aruw-mcb-project/src/aruwsrc/robot/sentry/turret/sentry_turret_major_world_relative_yaw_controller.hpp
@@ -98,7 +98,7 @@ public:
      * @see TurretControllerInterface for more details.
      * @param[in] desiredSetpoint The yaw desired setpoint in the world frame.
      */
-    void runController(const uint32_t dt, const WrappedFloat desiredSetpoint) final;
+    void runController(const float dt, const WrappedFloat desiredSetpoint) final;
 
     /// Sets the world frame yaw angle setpoint, refer to top level documentation for more details.
     void setSetpoint(WrappedFloat desiredSetpoint) final;
